reused common variables
