+++ beacon.go
@@ -17,32 +17,150 @@
 package beacon
 
 import (
+	"context"
+	"encoding/xml"
+	"errors"
+	"fmt"
 	"html/template"
 	"net/http"
 	"os"
+	"strconv"
+
+	"cloud.google.com/go/bigquery"
+	"google.golang.org/appengine"
+)
+
+type beaconConfig struct {
+	ApiVersion string
+	ProjectID  string
+	TableId    string
+}
+
+const (
+	apiVersionKey = "VERSION"
+	projectKey    = "GOOGLE_CLOUD_PROJECT"
+	bqTableKey    = "GOOGLE_BIGQUERY_TABLE"
 )
 
 var (
-	indexTemplate = template.Must(template.ParseFiles("index.xml"))
+	aboutTemplate = template.Must(template.ParseFiles("about.xml"))
+	config        = beaconConfig{
+		ApiVersion: os.Getenv(apiVersionKey),
+		ProjectID:  os.Getenv(projectKey),
+		TableId:    os.Getenv(bqTableKey),
+	}
 )
 
-type templateParams struct {
-	ApiVersion string
-	Datasets   string
-}
-
 func init() {
-	http.HandleFunc("/", handle)
+	http.HandleFunc("/", aboutBeacon)
+	http.HandleFunc("/query", query)
 }
 
-func handle(w http.ResponseWriter, r *http.Request) {
+func aboutBeacon(w http.ResponseWriter, r *http.Request) {
 	if r.Method != "GET" {
 		return
 	}
 	w.Header().Set("Content-Type", "application/xml")
-	indexTemplate.Execute(w, templateParams{
-		Datasets:   os.Getenv("TABLE"),
-		ApiVersion: os.Getenv("VERSION"),
-	})
+	aboutTemplate.Execute(w, config)
+}
 
+func query(w http.ResponseWriter, r *http.Request) {
+	if err := validateServerConfig(); err != nil {
+		http.Error(w, fmt.Sprintf("validating server configuration: %v", err), http.StatusInternalServerError)
+		return
+	}
+
+	refName, allele, coord, err := parseInput(r)
+	if err != nil {
+		http.Error(w, fmt.Sprintf("validating input: %v", err), http.StatusBadRequest)
+		return
+	}
+
+	ctx := appengine.NewContext(r)
+	exists, err := GenomeExists(ctx, refName, allele, coord)
+	if err != nil {
+		http.Error(w, fmt.Sprintf("computing result: %v", err), http.StatusInternalServerError)
+		return
+	}
+
+	if err := writeResponse(w, exists); err != nil {
+		http.Error(w, fmt.Sprintf("validating server configuration: %v", err), http.StatusInternalServerError)
+		return
+	}
+}
+
+func GenomeExists(ctx context.Context, refName string, allele string, coord int64) (bool, error) {
+	// Start is inclusive, End is exclusive.  Search exactly for coordinate.
+	query := fmt.Sprintf(`
+		SELECT count(v.reference_name) as count
+		FROM %s as v
+		WHERE reference_name='%s'
+			AND v.start <= %d AND %d < v.end
+	 	 	AND reference_bases='%s'
+		LIMIT 1`,
+		fmt.Sprintf("`%s`", config.TableId),
+		refName,
+		coord,
+		coord+1,
+		allele)
+
+	bqclient, err := bigquery.NewClient(ctx, config.ProjectID)
+	if err != nil {
+		return false, fmt.Errorf("creating bigquery client: %v", err)
+	}
+	it, err := bqclient.Query(query).Read(ctx)
+	if err != nil {
+		return false, fmt.Errorf("querying database: %v", err)
+	}
+
+	var result struct {
+		Count int
+	}
+	if err := it.Next(&result); err != nil {
+		return false, fmt.Errorf("reading query result: %v", err)
+	}
+	return result.Count > 0, nil
+}
+
+func validateServerConfig() error {
+	if config.ProjectID == "" {
+		return fmt.Errorf("%s must be specified", projectKey)
+	}
+	if config.TableId == "" {
+		return fmt.Errorf("%s must be specified", bqTableKey)
+	}
+	return nil
+}
+
+func parseInput(r *http.Request) (string, string, int64, error) {
+	refName := r.FormValue("chromosome")
+	if refName == "" {
+		return "", "", 0, errors.New("missing chromosome name")
+	}
+	allele := r.FormValue("allele")
+	if refName == "" {
+		return "", "", 0, errors.New("missing allele")
+	}
+	coord, err := strconv.ParseInt(r.FormValue("coordinate"), 10, 64)
+	if err != nil {
+		return "", "", 0, fmt.Errorf("parsing coordinate: %v", err)
+	}
+	return refName, allele, coord, nil
+}
+
+func writeResponse(w http.ResponseWriter, exists bool) error {
+	type beaconResponse struct {
+		XMLName struct{} `xml:"BEACONResponse"`
+		Exists  bool     `xml:"exists"`
+	}
+	var resp beaconResponse
+	resp.Exists = exists
+
+	w.Header().Set("Content-Type", "application/xml")
+	enc := xml.NewEncoder(w)
+	enc.Indent("", "  ")
+	if err := enc.Encode(resp); err != nil {
+		return fmt.Errorf("serializing response: %v", err)
+	}
+	return nil
 }
